job #9974 first prototype pass on partial parse
A lot of set up being done here. A lot of TODOs have been left in
the code in areas where assumptions are made. Still works based on
a preference being set. Only works if the change is in simple statements.
Changes that cross block boundaries or are made to if, while, and for
themselves are undefined at the moment.