job:#10010 Removed org.antrl.ui from the developer build dependencies
The version of org.antr.ui we are using has dependencies that require the
org.eclipse.core.runtime.compatibility plugin. That plugin was removed in
Eclipse Neon. This means we must update eclipse to use that. The only
reason we use that pluign is to get UI integration in eclipse, and we do not
really need that integration. We use it, occasionlly, for "Compile Antrl grammer",
but that s the only reason.